feat(core): extract utils/ into core/ kernel (Phase B)
Move logger, crypto, log-sanitizer, network-security, and
pino-elasticsearch-transport (+ colocated tests) to
backend/src/core/utils/. Keep pii-scrubber.ts in utils/ (domain-specific,
not kernel). Create re-export shims at original paths.

Part of #706
