+++ backend/src/core/utils/log-sanitizer.ts
@@ -1,0 +1,24 @@
+const SENSITIVE_KEYS = [
+  'password',
+  'secret',
+  'token',
+  'apikey',
+  'api_key',
+  'authorization',
+  'x-api-key',
+];
+
+export function sanitize(obj: unknown): unknown {
+  if (typeof obj !== 'object' || obj === null) return obj;
+  if (Array.isArray(obj)) return obj.map(sanitize);
+
+  const result: Record<string, unknown> = {};
+  for (const [key, value] of Object.entries(obj as Record<string, unknown>)) {
+    if (SENSITIVE_KEYS.some((s) => key.toLowerCase().includes(s))) {
+      result[key] = '[REDACTED]';
+    } else {
+      result[key] = sanitize(value);
+    }
+  }
+  return result;
+}
